refactor(app): extract per-setting prompt orchestrators (phase 4a)
Move 14 settings-prompt method triples (start_*_prompt / apply_* /
save_*_to_config) for line ending, encoding, language, theme,
keybinding map, cursor style, locale, ruler, plus the
reapply_all_overlays helper out of input.rs into a new
app/settings_prompts.rs sibling. ~852 lines.

These were prompt orchestrators living in input.rs by historical
accident — they're not key-dispatch code. Visibility tightened to
pub(super) so handle_action's match arms can still call them.

A future phase will collapse the extreme boilerplate (each triple is
~50 lines of near-identical "build suggestions, set prompt") into a
single SettingsPromptBuilder<T> generic over the option type.

input.rs goes from 3793 -> 2941 lines (-852). 2026 tests still pass.
